requirements.txt
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
audioop-lts==0.2.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
decorator==5.2.1
dotenv==0.9.9
executing==2.2.0
fastapi==0.115.12
ffmpy==0.5.0
filelock==3.18.0
fsspec==2025.3.2
gradio==5.23.3
gradio-client==1.8.0
groovy==0.1.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.1
idna==3.10
ipython==9.1.0
ipython-pygments-lexers==1.1.1
jedi==0.19.2
jinja2==3.1.6
jsonpickle==4.0.5
markdown-it-py==3.0.0
markupsafe==3.0.2
matplotlib-inline==0.1.7
mcp==1.6.0
mdurl==0.1.2
networkx==3.4.2
numpy==2.2.4
ollama==0.4.7
orjson==3.10.16
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
prompt-toolkit==3.0.50
ptyprocess==0.7.0
pure-eval==0.2.3
pydantic==2.11.2
pydantic-core==2.33.1
pydantic-settings==2.8.1
pydub==0.25.1
pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
pyvis==0.3.2
pyyaml==6.0.2
requests==2.32.3
rich==14.0.0
ruff==0.11.3
safehttpx==0.1.6
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sse-starlette==2.2.1
stack-data==0.6.3
starlette==0.46.1
tomlkit==0.13.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.15.2
typing-extensions==4.13.1
typing-inspection==0.4.0
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.34.0
wcwidth==0.2.13
websockets==15.0.1
